requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
playwright == 1.43.0
disposable-email-domains==0.0.98
dnspython==2.0.0
bs4==0.0.2
phonenumbers==8.13.32
pdfplumber==0.11.0
timeout_decorator==0.5.0